fix a syncing bug
Previous versions appended an "imported from..." message to the original
iCal event description to make it apparent that it wasn't a manually created
event. However this makes the iCal and GCal descriptions different! So
every single time the script ran it would think the iCal event had been
changed and needlessly update the GCal event again. It now uses the GCal
'source' field instead, filling it with a link to the original iCal feed.